backend/requirements.txt
Flask==3.0.0
pandas==2.2.2
Flask_Cors==4.0.0
tensorboard==2.15.0
protobuf==3.20.3
Flask-SocketIO==5.3.6
# openai==1.16.2
openai==1.30.3
python_dotenv==1.0.1
pyAgrum==1.13.0
IPython==8.23.0
tabulate==0.9.0
#dowhy==0.11.1 #needs to be installed with conda
#osqp #needs to be installed with conda